shifting logging funcs from main to utils
