+++ main.py
@@ -37,6 +37,7 @@ from src.utils.utils import (
     maybe_make_dir,
     parse_json_diff,
     sort_param_recursive,
+    select_relevant_print_args
 )
 import jsondiff
 
@@ -295,76 +296,6 @@ def evaluate_and_write(args, model, tasks, splits_to_write):
     results_tsv = os.path.join(args.exp_dir, "results.tsv")
     log.info("Writing results for split 'val' to %s", results_tsv)
     evaluate.write_results(val_results, results_tsv, run_name=run_name)
-
-
-def select_relevant_print_args(args):
-    """
-    Selects relevant arguments to print out. 
-    We select relevant arguments as the difference between defaults.conf and the experiment's 
-    configuration. 
-
-    Params
-    -----------
-    args: Params object
-
-    Returns
-    -----------
-    return_args: Params object with only relevant arguments
-    """
-    import pyhocon
-
-    exp_config_file = os.path.join(args.run_dir, "params.conf")
-    defaults_file = os.path.join(
-        os.path.dirname(os.path.realpath(__file__)) + "/config/defaults.conf"
-    )
-    exp_basedir = os.path.dirname(exp_config_file)
-    default_basedir = os.path.dirname(defaults_file)
-    fd = open(exp_config_file, "r")
-    exp_config_string = fd.read()
-    exp_config_string += "\n"
-    fd = open(defaults_file, "r")
-    default_config_string = fd.read()
-    default_config_string += "\n"
-    exp_config = dict(
-        pyhocon.ConfigFactory.parse_string(exp_config_string, basedir=exp_basedir).items()
-    )
-    default_config = dict(
-        pyhocon.ConfigFactory.parse_string(default_config_string, basedir=default_basedir).items()
-    )
-    sorted_exp_config = sort_param_recursive(exp_config)
-    sorted_defaults_config = sort_param_recursive(default_config)
-    diff_args = parse_json_diff(jsondiff.diff(sorted_defaults_config, sorted_exp_config))
-    diff_args = config.Params.clone(diff)
-    result_args = select_task_specific_args(args, diff_args)
-    return return_ags
-
-
-def select_task_specific_args(exp_args, diff_args):
-    """
-    A helper function that adds in task-specific parameters from the experiment 
-    configurations for tasks in pretrain_tasks and target_tasks.
-    """
-    exp_tasks = []
-    if diff_args.get("pretrain_tasks"):
-        exp_tasks = diff_args.pretrain_tasks.split(",")
-    if diff_args.get("target_tasks"):
-        exp_tasks += diff_args.target_tasks.split(",")
-    if len(exp_tasks) == 0:
-        return diff_args
-    for key, value in list(exp_args.as_dict().items()):
-        stripped_key = key.replace("_", " ")
-        stripped_key = stripped_key.replace("-", " ")
-        param_task = None
-        # For each parameter, identify the task the parameter relates to (if any)
-        for task in exp_tasks:
-            if task in stripped_key and (("edges" in stripped_key) == ("edges" in task)):
-                # special logic for edges since there are edge versions of various
-                # tasks.
-                param_task = task
-        # Add parameters that pertain to the experiment tasks
-        if param_task and param_task in exp_tasks:
-            diff_args[key] = value
-    return diff_args
 
 
 def initial_setup(args, cl_args):
